Changed every part of the Scene loader: Levels are now loaded via a Scene graph, every renderable entity is stored within a graph structure, each node complete with attached Entity and local transforms - compilable to world trans. by parsing the scene tree (or calling GetWorldTransform() per Node).
Also, the entire render pipeline got a cleaning out:
A Scene graph is passed to VisibleSet, along with a camera that can
determine which nodes are renderable (not fully operational yet).
The VisibleSet is then given to the Pipeline to be rendered, and each
RenderPass takes the VisibleSet to render from, depending on entities
being rendered.

A few classes have been removed, Graph.h, Plane, with view to removing
a few more with further commits.
Two new Mesh flags have been added, to determine whether the Mesh is
suitable for DEFERRED or FORWARD rendering paths, defaulting to DEFERRED.